task1: restructure files. Task2: during update of apointment, automatically jump to top of screen.
