--- src/App.css
@@ -1,5 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-}
-
