+++ src/Create150Appointments.js
@@ -2,7 +2,7 @@ import React from "react";
 import { useEffect, useRef } from "react";
 import { useSelector } from "react-redux";
 import { useDispatch } from "react-redux";
-import "./App.css";
+// import "./App.css";
 
 import clientsDentistCompanyBVT from "./dataInDentistAppWhenDentistAppStarts/clients"
 import dentistsDentistCompanyBVT from "./dataInDentistAppWhenDentistAppStarts/dentists"
@@ -34,7 +34,7 @@ const Create150Appointments = () => {
         The bigger the nr, the bigger the chance that the automatically generated appointment requires the precence of an assistant.
     */
     const CONFIGNROFASSISTANTS = 4; 
-    const CONFIGNROFRANDOMLYGENERATEDAPPOINTMENTS = 15;
+    const CONFIGNROFRANDOMLYGENERATEDAPPOINTMENTS = 150;
     const CONFIGNROFCLIENTS = 50;
     const CONFIGNROFDENTISTS = 4; 
     const CONFIGNROFDIFFERENTTREATMENTSASTHESKILLSOFADENTIST = 12;
