+++ src/components/app/App.js
@@ -3,22 +3,24 @@ import { ThemeProvider } from 'styled-components';
 import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 import { useSelector } from "react-redux";
 
-import Create150Appointments from "./Create150Appointments";
+
+
+import Create150Appointments from "../../Create150Appointments";
 // import {UseEffectHookCalls} from "./components/obsolete_fnCalls_winc_operations_witout_UI/UseEffectHookCalls";
-import Appointment from "./components/appointment/Appointment";
-import Assistant from "./components/assistant/Assistant";
-import Client from "./components/client/Client";
-import Dentist from "./components/dentist/Dentist";
-import {Calendar} from "./components/monthView/Calendar";
-import {SelectDayNrToDisplay} from "./components/dayView/SelectDayNrToDisplay";
+import Appointment from "../appointment/Appointment";
+import Assistant from "../assistant/Assistant";
+import Client from "../client/Client";
+import Dentist from "../dentist/Dentist";
+import {Calendar} from "../monthView/Calendar";
+import {SelectDayNrToDisplay} from "../dayView/SelectDayNrToDisplay";
 
-import {log} from './utils';
+import {log} from '../../utils';
 
-import {theme} from "./appStyleTheme"
+import {theme} from "../styles/appStyleTheme"
 // import "./App.css";
 import { AppStyled } from "./App.styled";
 import {NavigationStyled} from "./Navigation.styled";
-
+import {StyledNavLink} from "./NavLink.styled";
 
 const App = ()  => {
 
@@ -31,7 +33,6 @@ const App = ()  => {
   theme.pointerEvents = "none"; // works
   theme.pointerEvents = uiControlBehaviorDuringAppointmentUpdate.pointerEvents
 
-
   let appointmentsfromReduxToolkit = useSelector((state) => state.appointment)
   return(
     <>
@@ -42,24 +43,24 @@ const App = ()  => {
           <NavigationStyled>
             <nav className="navBar">
               <ul>  
-                <li>
+                <StyledNavLink>
                   <Link to="/">Calendar view</Link>
-                </li>
-                <li>
+                </StyledNavLink>
+                <StyledNavLink>
                   <Link to="/day">Day view</Link>
-                </li>
-                <li>
+                </StyledNavLink>
+                <StyledNavLink>
                   <Link to="/appointment">Appointment</Link>
-                </li> 
-                <li>
+                </StyledNavLink> 
+                <StyledNavLink>
                   <Link to="/assistant">Assistant</Link>
-                </li>
-                <li>
+                </StyledNavLink>
+                <StyledNavLink>
                   <Link to="/client">Client</Link>
-                </li>
-                <li>
+                </StyledNavLink>
+                <StyledNavLink>
                   <Link to="/dentist">Dentist</Link>
-                </li>
+                </StyledNavLink>
               </ul>
             </nav>
             <main>
@@ -86,7 +87,7 @@ const App = ()  => {
             </main>
           </NavigationStyled>
         </Router>
-        </AppStyled>
+      </AppStyled>
     </ThemeProvider>
   </>
  );
