+++ src/components/app/NavLink.styled.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const StyledNavLink = styled.li`
+  &:hover {
+    font-size: 2.04rem;
+  }
+`
