--- src/components/dayView/Day.css
@@ -1,73 +1,0 @@
-.appointmentInDayView {
-  /* list-style-type: none;
-  background-color: lightyellow;
-
-  border-top: 2px solid grey;
-  width: 92.8vw;
-  width: 76.8vw;
-  margin-bottom: 20px;
-
-  padding: 0.1rem;
-  border-top-style: solid;
-  border-width: 1px;
-  margin-top: 0.2rem;
-  margin-right: 2rem; */
-}
-
-.dayNrInDayView {
-  /* float: right;
-  font-weight: bold;
-  font-size: 1.1rem;
-
-  margin-top: 10px;
-  margin-right: 10px; */
-
-}
-
-
-
-.assistantInDayViewAppointment,
-.dentistInDayViewAppointment {
-  padding: 10px;
-  font-size: 1.1rem;
-  padding-left: 11vw;
-}
-
-.clientInDayView,
-.timeInDayView {
-  /* height: 30px;
-  padding: 10px; */
-}
-
-.clientInDayView {           
-  /* font-weight: bold;
-  background-color: deepskyblue;
-  font-size: 1.1rem;
-  padding-left: 7vw; */
-}
-
-
-.timeInDayView {
-  /* float: right;
-  background-color: aquamarine;
-  font-size: 1.1rem; */
-}
-
-
-
-.treatmentType {
-  /* float: right;
-  background-color: aquamarine;
-  font-size: 1.1rem;
-  padding: 6px;
-  margin-right: 20px;
-  margin-left: 20px;
-  margin-top: 20px; */
-}
-
-
-
-
-
-
-
