+++ src/components/styles/GlobalCss.styled.js
@@ -1,6 +1,6 @@
 import { createGlobalStyle } from 'styled-components'
 
-export default createGlobalStyle`
+const GlobalStyle =  createGlobalStyle`
  
     *, *::before, *::after {
         box-sizing: border-box;
@@ -15,8 +15,19 @@ export default createGlobalStyle`
       }
     
 
-    // body {
-    //     color: ${props => (props.darkMode ? 'white' : 'black')};
-    // }
+      body {
+        margin: 0;
+        font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+          'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+          sans-serif;
+        -webkit-font-smoothing: antialiased;
+        -moz-osx-font-smoothing: grayscale;
+      }
+      
+      code {
+        font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+          monospace;
+      }
 
 `
+export default GlobalStyle
