--- src/components/styles/GooglePlayStore.styled.js
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-export const StyledGooglePlayStore = styled.div`
-    margin: 1.25rem;
-
-    li {
-        margin-left: 1rem;
-    }
-    font-family: ${({ theme}) => theme.font };
-        
-    
-`
