--- src/components/styles/NavLink.styled.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components'
-
-export const StyledNavLink = styled.div`
-width: 100%;
-padding-left: 40%;
-font-size: 1.4rem;
-  &:hover {
-    font-size: 1.6rem;
-  }
-`
